# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: "Payment Order — BIAN Semantic API"
  description: >
    BIAN Service Domain 'Payment Order' (Process) on the 'Payment Order
    Procedure' control record — payment intake, validation, and routing to
    Payment Execution. States: RECEIVED → VALIDATED → SUBMITTED →
    COMPLETED | FAILED; REJECTED at validation; CANCELLED only before the
    execution hand-off.
  version: 0.2.0
servers:
  - url: /
  - url: /sd-payment-order
    description: through the platform gateway
paths:
...
  /v1/payment-order-procedure/{orderId}/control:
    put:
      operationId: cancelOrder
      summary: "Control — cancel (only before the execution hand-off)"
      parameters: [{ $ref: "#/components/parameters/OrderId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: { type: string, enum: [cancel] }
      responses:
        "200": { description: Order CANCELLED }
        "400": { $ref: "#/components/responses/Invalid" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/RuleViolation" }
  /v1/payment-order-procedure/{orderId}/execution-result:
    put:
      operationId: applyExecutionResult
      summary: "Outcome callback from Payment Execution (SUBMITTED orders only)"
      parameters: [{ $ref: "#/components/parameters/OrderId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [completed]
              properties:
                completed: { type: boolean }
                reason: { type: string }
      responses:
        "200": { description: Order COMPLETED or FAILED }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/RuleViolation" }
components:
  parameters:
    OrderId:
      name: orderId
      in: path
      required: true
      schema: { type: string }
  responses:
    NotFound:
      description: Unknown order
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RuleViolation:
      description: State forbids this (already submitted, terminal, not validated…)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Invalid:
      description: Bad input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        code: { type: string, example: ALREADY_SUBMITTED }
        message: { type: string }